fix(TimeBars): split up large file
